5.6 Update Level Building Area 4
Redoing area 4 and adding new limits and roofing, started to add
grayboxing for the end and beginning of that area. All areas 1 - 2 are
completed. With the exception of area 3 and 5.

Area 4 is currently focus of next sprint, with area 3 on the curtails of
the gray box.

Character model and area spawning fixed, with lighting rebuilt for
entire map.