9.5 The Issue Update
Fixing Lighting Areas and Problems with Maze 2
Textures have be added to as many objects as possible
Will try to finish up sound cues to less loud and a bit more random. So
that way the player is given more options - Modulation is now being
added.
After learning about sound cues and modulation tools inside of the sound
editor in unreal this will be the next focus and objects and textures
will be fixed up and added in process
Player Interface is going to default to pause men
Pause Menu button replaced exit/menu button from testing.
Removing any packaging issues.
Keeping Cost of level rendering down to a minimum.