10.0 The Pause Menu
Finally it is here. The End Screen, The Pause Menu and the functionality
of menus. I was working on trying to get these in and struggled to
figure out how to implement it.

Character Modes were added for each of these modes and some ambient
soundtracks were fixed to be better.

All areas have been vastly improved in terms of tile sets

Some areas are still suffering a tiny bit. But whille be dealt at a
later date.

Bugs : :-1: Small Bug with character screen causing issues. Delay not
triggering. Event Begin play is used too often.

Will have to add a loading screen to prevent players from seeing the
object screen.