10.2 Area 5 Rework
Area 5 Seems to be the biggest contention and problem I've had with
testing with users. Each area Needs to feel different in tone and still
match it.

For this area I want it to feel like a cave / throne room. The throne
room matches its theme, but not its tone that I originally set out to
do. For this we will reorganize the throne room. And have a good reason
as to why the player wants to go to there. The reason is just driving in
as to why they need to be there. So each area will now just have a key
that the player finds that unlocks the next area that all tie to a
single object that need to get. A set of objects and things that the
player has like a hub area. unlocking a gate.