10.7 Area 4 Update and Area 2 Material Update
Added Materials to Area 2
Increased the size of the walls of area 4 to be a bit more in theme of
unknown, so it feels like your completely alone

Added junk to area 5 to prevent the player from doing certain things

Bug - Editor for some reason spawns objects thousands of blocks away
from where intially placed?

Might need to check to see why that is happening

Objects are slower but not due to infrequencies

As this is nearing completion thankfully most of these objects are not
needed as they have already been placed down.

TO DO : Finish up sounds in final area, textures, materials, triggers,
and the banner.