13.1 Area 01 - Area 03 Update
Area 01 and Area 03 update

Redid all of the materials to be less emissive will continue to work on
the materials to be less  (eye hurting)

Cannot get rid of eye effect due to environmental need.

Area 01 rebuilt and Level Depths Maze has been completely rebuilt and
redesigned to allow for the player to get to somewhere.

Area 4 and 5 are now on the bucket list to be redone.

Sounds for God creature are to be added.