13.3 Disappearing and Reappearing Text, Cave Update
Fixed issues with text not appearing exactly like that I wanted it to

Added in blue print text

Removed Auto Exposure Histogram

Created a Better lighting system

Also added roof top and fixed lighting issues on map.

Going to be adding lights to the areas to give a bit more contrast to
all the areas.

Going to be editting the fog effect to be more interesting and less
PURPLE

light intensity of the sun will be turned down just a tad.

Fixing Hellgates to be less obtrusive and adding in lighting to get rid
of emissive flares.

Added in Sky Light

Added in more lights to various areas to make it less extreme.