14.5 Physical Materials and footsteps
Cleaned up code added in new cues that are better for detection system.

Footsteps are a bit off.

And the chain is quite annoying will probably remove or add a chain
sound that doesn't rattle so it will be different each other time.

Physical materials are being changed through out.