14.6 Camera Shake / Bob
Camera shakes whenever you are called by the god will be creating it to
be a bit more random with the shake but just random enough that it is
too varied.

Will also be looking into creating a new sound effect to make it better
and a bit more immersive so that way people cannot tell the difference
between all the sounds.

Will going in this weekend to edit these sounds.

Area 01 will need to be looked at for design